// src/data/wiring.generated.ts
export const connectorPinouts: ConnectorPinout[] = [
// ...
  {
    "id": "lamp-assembly-left-front-c1-connector-details",
    "meta": {
      "Harness Family": "Dash",
      "Color": "BLACK",
      "Gender": "FEMALE",
      "Part Number": "33472-0601",
      "Part Description": "Connector No:",
      "Connector No": "D6605A",
      "Option": "Standard",
      "Repair Kit Part Number": "68225123AA",
      "Cavity Count": "6",
      "name": "LAMP-ASSEMBLY-LEFT FRONT C1"
    },
// ...
  },
// ...
];